refactor(domain): AttendancePolicy 클래스 생성
출석 객체의 행위가 아닌, 별도의 독립된 출석 정책이라 생각되는 메서드 AttendancePolicy 클래스로 분리
