+++ src/main/java/domain/Attendance.java
@@ -4,14 +4,9 @@ import java.time.DayOfWeek;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
-import java.util.List;
 import java.util.Objects;
 
 public class Attendance {
-    private static final LocalTime CAMPUS_START_TIME = LocalTime.of(8, 0);
-    private static final LocalTime CAMPUS_END_TIME = LocalTime.of(23, 0);
-    private static final List<Integer> HOLIDAY = List.of(25);
-
     private final LocalDateTime dateAndTime;
 
     public Attendance(LocalDateTime dateAndTime) {
@@ -20,17 +15,6 @@ public class Attendance {
 
     public Attendance(LocalDate date, LocalTime time) {
         this.dateAndTime = LocalDateTime.of(date, time);
-    }
-
-    public static boolean isInvalidTime(LocalTime time) {
-        return time.isBefore(CAMPUS_START_TIME) || time.isAfter(CAMPUS_END_TIME);
-    }
-
-    public static boolean isHoliday(LocalDate date) {
-        DayOfWeek dayOfWeek = date.getDayOfWeek();
-        return dayOfWeek == DayOfWeek.SATURDAY
-                || dayOfWeek == DayOfWeek.SUNDAY
-                || HOLIDAY.contains(date.getDayOfMonth());
     }
 
     public LocalDate getDate() {
